# bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "FinanceTradingApi, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/api/Portfolios": {
      "get": {
        "tags": [
          "Portfolios"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Portfolio"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Portfolios"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Portfolio"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Portfolio"
            }
          }
        }
      }
    },
    "/api/Portfolios/{id}": {
      "get": {
        "tags": [
          "Portfolios"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Portfolio"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Portfolios"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Portfolio"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Stocks": {
      "get": {
        "tags": [
          "Stocks"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Stock"
              }
            }
          }
        }
      }
    },
    "/api/Stocks/{symbol}": {
      "get": {
        "tags": [
          "Stocks"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "symbol",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Stock"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Portfolio": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "userId": {
          "type": "string"
        },
        "stocks": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Stock"
          }
        },
        "totalValue": {
          "format": "double",
          "type": "number"
        }
      }
    },
    "Stock": {
      "type": "object",
      "properties": {
        "symbol": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "price": {
          "format": "double",
          "type": "number"
        },
        "change": {
          "format": "double",
          "type": "number"
        },
        "percentChange": {
          "format": "double",
          "type": "number"
        }
      }
    }
  }
}